Move VRG VolSync functions to a separate file
