+++ config/crd/bases/ramendr.openshift.io_volumereplicationgroups.yaml
@@ -288,6 +288,10 @@ spec:
                             name:
                               description: Name of the VolRep/PVC resource
                               type: string
+                            protectedByVolSync:
+                              description: VolSyncPVC can be used to denote whether
+                                this PVC is protected by VolSync. Defaults to "false".
+                              type: boolean
                             resources:
                               description: Resources set in the claim to be replicated
                               properties:
@@ -319,10 +323,6 @@ spec:
                               description: Name of the StorageClass required by the
                                 claim.
                               type: string
-                            volSyncPVC:
-                              description: VolSyncPVC can be used to denote whether
-                                this PVC is protected by VolSync. Defaults to "false".
-                              type: boolean
                           type: object
                         sshKeys:
                           description: sshKeys is the name of a Secret that contains
@@ -531,6 +531,10 @@ spec:
                     name:
                       description: Name of the VolRep/PVC resource
                       type: string
+                    protectedByVolSync:
+                      description: VolSyncPVC can be used to denote whether this PVC
+                        is protected by VolSync. Defaults to "false".
+                      type: boolean
                     resources:
                       description: Resources set in the claim to be replicated
                       properties:
@@ -561,10 +565,6 @@ spec:
                     storageClassName:
                       description: Name of the StorageClass required by the claim.
                       type: string
-                    volSyncPVC:
-                      description: VolSyncPVC can be used to denote whether this PVC
-                        is protected by VolSync. Defaults to "false".
-                      type: boolean
                   type: object
                 type: array
               state:
